Examples of a few possible extra validators
