--- test/global.d.ts
@@ -1,2 +1,0 @@
-
-declare const require: any
